+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Enhancements
 - GP2-3767 - EP pdf export template rework
+- GP2-3776 - EP Textarea autosize
 - GP2-3436 - Add created date to EP overview
 - GP2-3190 - WTE content changes
 - GP2-2478 - EP Business objectives content update